+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Era.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -12,9 +13,15 @@ import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
 import Cardano.Ledger.Babbage
 import Cardano.Ledger.Babbage.Core
 import Cardano.Ledger.Plutus (Language (..))
+import Paths_cardano_ledger_babbage (getDataFileName)
 import Test.Cardano.Ledger.Alonzo.Era
 import Test.Cardano.Ledger.Babbage.Arbitrary ()
 import Test.Cardano.Ledger.Babbage.Binary.Annotator ()
+import Test.Cardano.Ledger.Babbage.Examples (
+  exampleBabbageOnwardsEraPParams,
+  exampleBabbagePParamsUpdate,
+  exampleBabbageTx,
+ )
 import Test.Cardano.Ledger.Babbage.TreeDiff ()
 import Test.Cardano.Ledger.Plutus (zeroTestingCostModels)
 
@@ -27,10 +34,33 @@ class
   BabbageEraTest era
 
 instance EraTest BabbageEra where
+  type
+    EraRulesWithFailures BabbageEra =
+      '[ "BBODY"
+       , "DELEG"
+       , "DELEGS"
+       , "DELPL"
+       , "LEDGER"
+       , "LEDGERS"
+       , "POOL"
+       , "PPUP"
+       , "UTXO"
+       , "UTXOS"
+       , "UTXOW"
+       ]
   zeroCostModels = zeroTestingCostModels [PlutusV1 .. PlutusV2]
+
   mkTestAccountState = mkShelleyTestAccountState
 
   accountsFromAccountsMap = shelleyAccountsFromAccountsMap
+
+  mkEraFullPath = getDataFileName
+
+  exampleTx = exampleBabbageTx
+
+  examplePParams = exampleBabbageOnwardsEraPParams
+
+  examplePParamsUpdate = exampleBabbagePParamsUpdate
 
 instance ShelleyEraTest BabbageEra