+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Imp.hs
@@ -2,17 +2,20 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Babbage.Imp (spec, babbageEraSpecificSpec) where
+module Test.Cardano.Ledger.Babbage.Imp (
+  spec,
+  Alonzo.shelleyToBabbageSpec,
+  Alonzo.alonzoToConwaySpec,
+  babbageOnlySpec,
+) where
 
 import Cardano.Ledger.Babbage.Core
 import Cardano.Ledger.Babbage.State
-import Cardano.Ledger.Shelley.Rules
+import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import qualified Test.Cardano.Ledger.Alonzo.Imp as Alonzo
 import Test.Cardano.Ledger.Alonzo.ImpTest
 import qualified Test.Cardano.Ledger.Babbage.Imp.PoolSpec as POOL
@@ -24,7 +27,7 @@ import Test.Cardano.Ledger.Imp.Common
 
 spec ::
   ( BabbageEraImp era
-  , Event (EraRule "RUPD" era) ~ RupdEvent
+  , Shelley.Event (EraRule "RUPD" era) ~ Shelley.RupdEvent
   ) =>
   proxy era ->
   Spec
@@ -35,13 +38,13 @@ spec era = do
     UTXOW.spec
     UTXOS.spec
 
-babbageEraSpecificSpec ::
+babbageOnlySpec ::
   ( BabbageEraImp era
   , ShelleyEraAccounts era
-  , Event (EraRule "NEWEPOCH" era) ~ ShelleyNewEpochEvent era
+  , Shelley.Event (EraRule "NEWEPOCH" era) ~ Shelley.ShelleyNewEpochEvent era
   ) =>
   proxy era ->
   Spec
-babbageEraSpecificSpec era = do
+babbageOnlySpec era = do
   describe "BabbageEra Specific" $ withImpInitEachEraVersion era $ do
     POOL.babbageEraSpecificSpec