+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/Era.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -9,10 +11,16 @@ module Test.Cardano.Ledger.Mary.Era (
 import Cardano.Ledger.Mary
 import Cardano.Ledger.Mary.Core
 import Cardano.Ledger.Plutus (emptyCostModels)
+import Paths_cardano_ledger_mary (getDataFileName)
 import Test.Cardano.Ledger.Allegra.Era
 import Test.Cardano.Ledger.Mary.Arbitrary ()
 import Test.Cardano.Ledger.Mary.Binary.Annotator ()
+import Test.Cardano.Ledger.Mary.Examples (exampleMaryTx)
 import Test.Cardano.Ledger.Mary.TreeDiff ()
+import Test.Cardano.Ledger.Shelley.Examples (
+  exampleShelleyPParams,
+  exampleShelleyPParamsUpdate,
+ )
 
 class
   ( AllegraEraTest era
@@ -21,11 +29,33 @@ class
   MaryEraTest era
 
 instance EraTest MaryEra where
+  type
+    EraRulesWithFailures MaryEra =
+      '[ "BBODY"
+       , "DELEG"
+       , "DELEGS"
+       , "DELPL"
+       , "LEDGER"
+       , "LEDGERS"
+       , "POOL"
+       , "PPUP"
+       , "UTXO"
+       , "UTXOW"
+       ]
+
   zeroCostModels = emptyCostModels
 
   mkTestAccountState = mkShelleyTestAccountState
 
   accountsFromAccountsMap = shelleyAccountsFromAccountsMap
+
+  mkEraFullPath = getDataFileName
+
+  exampleTx = exampleMaryTx
+
+  examplePParams = exampleShelleyPParams
+
+  examplePParamsUpdate = exampleShelleyPParamsUpdate
 
 instance ShelleyEraTest MaryEra